refactored the code and moved it into separate files
