+++ src_package_arduino/include/DHT_handler.h
@@ -2,21 +2,8 @@
 #define DHT_HANDLER_H
 
 #include "DHT.h"
+#include "sensor_package_manager.h"
 #include <Arduino.h>
-
-// Queue size constant
-constexpr size_t QUEUE_SIZE = 100;
-
-// Sensor packet structure
-typedef struct
-{
-    uint8_t sensor_id;
-    uint32_t sensor_timestamp;
-    float temperature;
-    float humidity;
-    uint32_t server_package_id;
-    uint16_t package_sequence_number;
-} SensorPacket;
 
 // DHT sensor object
 extern DHT dht;
@@ -25,26 +12,8 @@ extern DHT dht;
 extern const uint8_t DHT_PIN;
 extern const uint8_t DHT_TYPE;
 
-// Sensor variables
-extern uint8_t sensor_id;
-extern float temperature;
-extern float humidity;
-extern uint32_t server_package_id;
-extern uint16_t package_sequence_number;
-
-// Circular buffer for sensor packets
-extern SensorPacket buffer[QUEUE_SIZE];
-extern size_t queue_head;
-extern size_t queue_tail;
-extern size_t queue_count;
-
 // Function declarations
 void initializeDHT();
 void readDHT(SensorPacket &packet);
-SensorPacket assembleSensorPacket();
-bool addPacketToBuffer(const SensorPacket &packet);
-bool getPacketFromBuffer(SensorPacket &packet);
-void flushBuffer();
-uint32_t getTimestamp();
 
 #endif
